pyproject.toml
[tool.poetry]

name = "jsonmask"
version = "0.0"
description = "Implements the Google Partial Response protocol in Python"

license = "MIT"

authors = ["Craig Labenz <craig.labenz@gmail.com>"]

readme = "README.md"

documentation = "https://github.com/zapier/jsonmask/blob/master/README.md"
repository = "https://github.com/zapier/jsonmask"

keywords = [
    "jsonmask",
    "json",
    "google partial response",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 2",
    "Programming Language :: Python :: 2.7",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
]

[tool.poetry.dependencies]

python = "~2.7 || ^3.6"

[tool.poetry.dev-dependencies]


# Linters
isort = "*"
pylint = "*"
pycodestyle = "*"
pydocstyle = "*"
snowballstemmer = "*" # missing dependency for pydocstyle

# Testing
pytest = "^3.3"
pytest-describe = "*"
pytest-expecter = "*"
pytest-random = "*"
pytest-cov = "*"
freezegun = "*"
minilog = "~0.4"

# Reports
coverage-space = "*"

# Documentation
mkdocs = "<1.0"
pygments = "*"

# Build
pyinstaller = "*"

# Tooling
sniffer = "*"
MacFSEvents = { version = "*", platform = "darwin" }
pync = { version = "*", platform = "darwin" }
